src/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
twelvedata
newsapi-python
google-genai
